Add more classes to adhere to OOP principles.
Most of the program logic was stored in BabyGronk class, which
did not adhere much to OOP style of coding.

Parser is introduced to parse user input and produce Instruction
that has the correct command as well as args formatted, otherwise,
an InvalidInputException will be thrown.

TaskList is created to store tasks and handle add and delete operations.

Ui class handles the printing of strings to stdout as well as
exiting the program.