Refactor code
Code was a little hard to read and logic was
messy.

Broke down some functions into parts and created
an enum InvalidCommands for
InvalidInputException.

Importantly, created functions such as
isMarkCommand, isDeleteCommand, isTaskCommand
in Instruction class so that it can be used by
Parser and BabyGronk class statically. Might
need to change this later.